apps/ai-service/pyproject.toml
[project]
name = "kerjacus-ai-service"
version = "0.1.0"
description = "KerjaCUS! AI Service - LLM gateway, CV parsing, talent matching"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.42.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "instructor>=1.14.0",
    "openai>=2.29.0",
    "docling>=2.81.0",
    "catboost>=1.2.10",
    "scikit-learn>=1.8.0",
    "psycopg[binary]>=3.3.0",
    "sentence-transformers>=5.3.0",
    "mlflow>=3.10.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "torch>=2.10.0",
    "pytest-cov>=7.1.0",
]

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-bdd>=8.1.0",
    "schemathesis>=4.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
